converro_backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.5.2
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
autopep8==1.6.0
axios==0.4.0
azure-core==1.30.2
Babel==2.10.3
bcc==0.29.1
beautifulsoup4==4.13.4
billiard==4.2.1
blinker==1.7.0
boto3==1.34.162
botocore==1.34.162
Brlapi==0.8.5
bs4==0.0.2
celery==5.4.0
certifi==2024.2.2
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloud-init==24.4.1
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
contourpy==1.3.2
cors==1.0.1
cryptography==43.0.3
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.3.2
defer==1.0.6
Deprecated==1.2.15
diff-match-patch==20241021
distro==1.9.0
distro-info==1.7+build1
dj-database-url==0.5.0
Django==4.2.16
django-celery-results==2.5.1
django-cors-headers==4.7.0
django-debug-toolbar==4.3.0
django-environ==0.8.1
django-import-export==4.1.1
django-storages==1.14.4
djangorestframework==3.14.0
dnspython==2.6.1
drf-yasg==1.21.8
et_xmlfile==2.0.0
exceptiongroup==1.2.2
filelock==3.16.1
fonttools==4.58.4
frozenlist==1.5.0
fsspec==2024.10.0
future==1.0.0
gevent==25.5.1
greenlet==3.2.3
gunicorn==23.0.0
gyp==0.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.20.4
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
inflection==0.5.1
Jinja2==3.1.4
jiter==0.8.0
jmespath==1.0.1
joblib==1.4.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
kiwisolver==1.4.8
kombu==5.4.2
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.29.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
neo4j==5.26.0
netaddr==0.8.0
netifaces==0.11.0
networkx==3.1
nltk==3.8.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
olefile==0.46
openai==1.55.1
opencv-python==4.9.0.80
openpyxl==3.1.5
packaging==24.2
pandas==2.0.3
pdf2image==1.17.0
pexpect==4.9.0
pikepdf==9.2.1
pillow==10.2.0
prompt_toolkit==3.0.51
propcache==0.3.2
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pycairo==1.25.1
pycodestyle==2.8.0
pycparser==2.22
pycups==2.0.1
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.17.2
PyGObject==3.48.2
PyJWT==2.7.0
pymongo==4.6.3
PyMuPDF==1.24.11
PyMySQL==1.1.1
pyotp==2.9.0
pyparsing==3.1.1
PyPDF2==3.0.1
pyrsistent==0.20.0
pyserial==3.5
PySocks==1.7.1
pytesseract==0.3.13
python-apt==2.7.7+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.49+ubuntu2
python-decouple==3.8
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.2
pyzbar==0.1.9
qrcode==8.0
razorpay==1.4.2
redis==5.2.0
regex==2023.12.25
requests==2.31.0
requests-file==2.1.0
rich==13.7.1
s3transfer==0.10.4
safetensors==0.4.5
scikit-learn==1.6.1
scipy==1.10.1
sentence-transformers==2.6.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.4.2
sympy==1.13.3
systemd-python==235
tablib==3.5.0
threadpoolctl==3.4.0
tldextract==5.3.0
tokenizers==0.15.2
toml==0.10.2
torch==2.2.2
tqdm==4.67.1
transformers==4.39.3
triton==2.2.0
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
uritemplate==4.2.0
urllib3==2.0.7
vine==5.1.0
wadllib==1.3.6
wcwidth==0.2.13
whitenoise==5.3.0
wrapt==1.17.0
xdg==5
xkit==0.0.0
xlwt==1.3.0
yarl==1.15.2
zope.event==5.1
zope.interface==7.2
